refactor: rename to prettier, robust handling, add truncation)
