+++ src/interactions/index.ts
@@ -4,4 +4,4 @@ export * from "./context/intentsLookupContext.js";
 export * from "./slash/userinfo.js";
 export * from "./slash/bitfieldLookup.js";
 export * from "./slash/snowflakeInfo.js";
-export * from "./context/formatContext.js";
+export * from "./context/prettierContext.js";
